+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+
+- Use `phpstan` as a dev dependency to detect bugs
+
+### Added
+
+- Added `responseFactory` option to `__construct`
+
 ## [1.0.0] - 2018-09-23
 
 ### Added